Put priority on AJAX calls so RBP initializes properly
Without this commit, any AJAX calls result in the RBP plugin not
initializing. This causes huge problems, especially if you do not
have a default, non role based price.
One example of a problem would be the check for the cart session.
It checks to make sure the cart is still valid on each request, and
it uses the price to make sure the item is purchasable. In the case
that an AJAX request is made, we need the frontend plugin loaded
so that is_price will return proper prices and not see all the
items as non purchaseable.